v0fix: createCommand now properly types execute
If Shortcut is extended to allow different some property (e.g. args) that
differs depending on the command name, creating a command will now correctly
type the parameters of the execute function.

It doesn't also type the returned Command because it gets annoying to have such
an exact type since it's not assignable to plain Command (because the execute
function no longer matches).